alien symphony
I ultimately decided to drop the nav bar so that I could keep the linear gradient, and because the detriment of crowding the space on mobile outweighed the benefit of navigating through a list of only 10 items slightly faster. I also ultimately kept the bullet points, so that readers can clearly differentiate between song names if a song title takes multiple lines.